condvs.sh
this file is used  as  shortcut to the  "MMDVM_Bridge tune" command script of DVSwitch.  This can be used as a command line, crontab, and used in the rpt.conf, and allmon3 web.ini as a bookmark. 
I created a   directory called  "/local"  where  I keep all my custom files, for ease of use.
Don't forget to    sudo chmod +x to make it executable